Major Enhancements!
RX tests are now fully implemented for the APX line of radios, including getting BER readings from the radio to test the receiver. Groundwork has been done to allow this to be ported over other lines of radios in the future.

While working on this, I also uncovered a timing issue that would sometime cause the application to hang while doing power tests, delays were updated to prevent this bug.

This has been fully tested with a Aeroflex ifr2975, but more work will likely be required to get other instruments going.