The R2670 is now supported!
In order to make this possible, a couple minor tweaks to the way the instuments methods were implemented was necessairy. The syntax for the R2670 is not as flexible as other intruments, some actions had to be grouped into one method (frequency and power had to be specified together for example).

The R2670 has limited programability with P25. The audio settings (to set the P25 patterns) have to be entered via commands emulating key presses as there are no commands to change this setting. The position of the cursor needs to be in it's default location for the test to work. If it is not, it is not fatal, but RX BER tests will report an error.