The R2670 is now supported!
~~With Recommended changes in PR#13 made~~

In order to make this possible, a couple minor tweaks to the way the instruments methods were implemented was necessary. The syntax for the R2670 is not as flexible as other instruments, some actions had to be grouped into one method (frequency and power had to be specified together for example).

The R2670 has limited programmability with P25. The audio settings (to set the P25 patterns) have to be entered via commands emulating key presses as there are no commands to change this setting. The position of the cursor needs to be in it's default location for the test to work. If it is not, it is not fatal, but RX BER tests will report an error.

Also, the R2670 offers no way of recovering the received P25 BER programmatically (for a radio TX test). Because of this, the TX:BER tests will not work for the R2670. In a future version of OAB-ng, we should hide the tests that are not relevant for a given instrument... this is not implemented here as this is out of scope for this commit.

This was tested on a IFR2975 and R2670A with the P25 Slice and P25 options enabled.